Name the types callers actually push in the MD printf
marsdev's string.c is written for its own Makefile, where -mshort makes int 16
bits and va_arg(args, int16_t) matches what a caller pushes. game/ deliberately
leaves -mshort off, so default argument promotion sends every integer through as
a 32 bit int and that va_arg names a type nothing ever passes.

It survives one argument and goes wrong with three. Adding a single %d to the
debug line was enough to leave Sonic falling through a stage that had been fine
a build earlier, with nothing in the diff that looked like it could reach
physics. Three runs of the reverted build were stable, a longer literal string
with the same arguments was fine, and three arguments printed with %X were fine,
which leaves the argument types and nothing else.

va_arg now names int and unsigned int, and the conversion locals are as wide as
the values they have to hold. The exact five argument format that broke now runs
with the stage intact. The copies under blitbench/ and cdbench/ are only ever
reached from -mshort builds and are correct where they sit.

Written up in section 6 with the tell: a change touching only a debug string
breaking something that is not the debug string.

+++ game/md_src/string.c
@@ -31,10 +31,10 @@ static uint16_t skip_atoi(const char **s)
 static uint16_t vsprintf(char *buf, const char *fmt, va_list args)
 {
     char tmp_buffer[12];
-    int16_t i;
+    int32_t i;
     int16_t len;
-    int16_t *ip;
-    uint16_t num;
+    int *ip;
+    uint32_t num;
     char *s;
     const char *hexchars;
     char *str;
@@ -90,7 +90,7 @@ repeat:
         {
             ++fmt;
             // it's the next argument
-            field_width = va_arg(args, int16_t);
+            field_width = va_arg(args, int);
 
             if (field_width < 0)
             {
@@ -109,7 +109,7 @@ repeat:
             {
                 ++fmt;
                 // it's the next argument
-                precision = va_arg(args, int16_t);
+                precision = va_arg(args, int);
             }
 
             if (precision < 0)
@@ -129,7 +129,7 @@ repeat:
                     while(--field_width > 0)
                         *str++ = ' ';
 
-                *str++ = (unsigned char) va_arg(args, int16_t);
+                *str++ = (unsigned char) va_arg(args, int);
 
                 while(--field_width > 0)
                     *str++ = ' ';
@@ -176,7 +176,7 @@ repeat:
 hexa_conv:
                 s = &tmp_buffer[12];
                 *--s = 0;
-                num = va_arg(args, uint16_t);
+                num = va_arg(args, unsigned int);
 
                 if (!num)
                     *--s = '0';
@@ -192,14 +192,14 @@ hexa_conv:
                 break;
 
             case 'n':
-                ip = va_arg(args, int16_t*);
+                ip = va_arg(args, int *);
                 *ip = (str - buf);
                 continue;
 
             case 'u':
                 s = &tmp_buffer[12];
                 *--s = 0;
-                num = va_arg(args, uint16_t);
+                num = va_arg(args, unsigned int);
 
                 if (!num)
                     *--s = '0';
@@ -218,7 +218,7 @@ hexa_conv:
             case 'i':
                 s = &tmp_buffer[12];
                 *--s = 0;
-                i = va_arg(args, int16_t);
+                i = va_arg(args, int);
 
                 if (!i)
                     *--s = '0';
